[level 1] Title: 크기가 작은 부분문자열, Time: 11.42 ms, Memory: 73.8 MB -BaekjoonHub

+++ 프로그래머스/1/147355. 크기가 작은 부분문자열/크기가 작은 부분문자열.java
@@ -22,6 +22,6 @@ class Solution {
             }
         }
         
-        return (int) cnt;
+        return cnt;
     }
 }
